+++ src/renderer/src/hooks/useParsedContent.ts
@@ -2,12 +2,18 @@ import { useMemo } from 'react';
 
 import { parseMessageContent } from '@/components/message/utils';
 
-export const useParsedContent = (baseDir: string, content: string | null | undefined, allFiles: string[], renderMarkdown?: boolean) => {
+export const useParsedContent = (
+  baseDir: string,
+  content: string | null | undefined,
+  allFiles: string[],
+  renderMarkdown?: boolean,
+  renderThinking?: boolean,
+) => {
   return useMemo(() => {
     if (!content) {
       return null;
     }
-    return parseMessageContent(baseDir, content, allFiles, renderMarkdown);
+    return parseMessageContent(baseDir, content, allFiles, renderMarkdown, renderThinking);
     // we use allFiles.length to re-evaluate if the array content might have changed
     // even if the array reference itself hasn't.
     // eslint-disable-next-line react-hooks/exhaustive-deps